Import libltc-1.2.0 as local/libltc.
Linear (or Longitudinal) Timecode (LTC) is an encoding of SMPTE timecode data
as a Manchester-Biphase encoded audio signal. The audio signal is commonly
recorded on a VTR track or other storage media.

libltc provides functionality to encode and decode LTC audio from/to SMPTE or
EBU timecode, including SMPTE date support.

libltc is the successor of libltcsmpte.