Import root-5.26nb17 as local/root.
The ROOT system provides a set of OO frameworks with all the functionality
needed to handle and analyse large amounts of data in a very efficient way.
Having the data defined as a set of objects, specialised storage methods are
used to get direct access to the separate attributes of the selected objects,
without having to touch the bulk of the data. Included are histograming methods
in 1, 2 and 3 dimensions, curve fitting, function evaluation, minimisation,
graphics and visualization classes to allow the easy setup of an analysis
system that can query and process the data interactively or in batch mode.

Thanks to the builtin CINT C++ interpreter the command language, the scripting,
or macro, language and the programming language are all C++. The interpreter
allows for fast prototyping of the macros since it removes the time consuming
compile/link cycle. It also provides a good environment to learn C++. If more
performance is needed the interactively developed macros can be compiled using
a C++ compiler.