Import 8cc-0nb20161001 as local/8cc.
8cc is a compiler for the C programming language. It's intended to support all
C11 language features while keeping the code as small and simple as possible.

The compiler is able to compile itself. You can see its code both as an
implementation of the C language and as an example of what this compiler is
able to compile.

8cc's source code is carefully written to be as concise and easy-to-read as
possible, so that the source code becomes good study material to learn about
various techniques used in compilers. You may find the lexer, the preprocessor
and the parser are already useful to learn how C source code is processed at
each stage.